+++ pkg/security/ebpf/kernel/kernel_bpf.go
@@ -18,7 +18,7 @@ import (
 	"github.com/cilium/ebpf/link"
 	"golang.org/x/sys/unix"
 
-	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
+	ddbtf "github.com/DataDog/datadog-agent/pkg/ebpf/btf"
 )
 
 // HaveMmapableMaps returns whether the kernel supports mmapable maps.
@@ -202,7 +202,7 @@ func (k *Version) HaveFentryNoDuplicatedWeakSymbols() bool {
 
 // SupportCORE returns is CORE is supported
 func (k *Version) SupportCORE() bool {
-	_, err := ddebpf.GetKernelSpec()
+	_, err := ddbtf.GetKernelSpec()
 	return err == nil
 }
